requirements.txt
requests==2.8.1
pycryptodome==3.3.1